Migrate more scheduled queries endpoints
